[Chore] chore query

+++ backend/src/main/java/com/hanati/team1/src/customers/models/GetMyInfoRes.java
@@ -3,7 +3,7 @@ package com.hanati.team1.src.customers.models;
 public record GetMyInfoRes(
 	String name,
 	String profile,
-	int totalAsset,
-	int totalCntToken
+	long totalAsset,
+	long totalCntToken
 ) {
 }
